cleaup(src): split backlight module in plugins.
